+++ ui/components/providers/workflow/forms/select-credentials-type/aws/radio-group-aws-via-credentials-type-form.tsx
@@ -59,7 +59,7 @@ export const RadioGroupAWSViaCredentialsTypeForm = ({
             </div>
           </RadioGroup>
           {errorMessage && (
-            <FormMessage className="text-system-error dark:text-system-error">
+            <FormMessage className="text-text-error">
               {errorMessage}
             </FormMessage>
           )}